feat: Rename JournalAbbreviationRepository to AbbreviationRepository (#30)
Closes #30
